# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Karuna Dham Foundation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /volunteers:
    get:
      operationId: getVolunteers
      tags: [volunteers]
      summary: List volunteers (admin)
      responses:
        "200":
          description: Volunteers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Volunteer"
    post:
      operationId: registerVolunteer
      tags: [volunteers]
      summary: Register as a volunteer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VolunteerInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Volunteer"
  /volunteers/me:
    get:
      operationId: getMyVolunteerProfile
      tags: [volunteers]
      summary: Get current user's volunteer profile
      responses:
        "200":
          description: Volunteer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Volunteer"
        "404":
          description: Not a volunteer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Contact
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Volunteer:
      type: object
      required: [id, name, email, skills, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        availability:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        status:
          type: string
        hoursContributed:
          type: integer
        createdAt:
          type: string
    VolunteerInput:
      type: object
      required: [name, email, skills]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        skills:
          type: array
          items:
            type: string
        availability:
          type: string
        location:
          type: string
        bio:
          type: string
...